backend (copy)/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
distro==1.9.0
dotenv==0.9.9
easyocr==1.7.2
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.93.0
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.16.0
shapely==2.1.1
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
